Switch from golang.org/x/net/websocket to github.com/gorilla/websocket
The Gorilla WebSocket library is more flexible and gives us
control of the underlying TCP connection.

This commit also fixes a bug in the previous implementation.
Previously, we were opening a persistant connection to the
server. Due to this, requests were getting mixed up resulting
in errors. We now open a new connection per request.