Implement Authentication
This commit implements SASL authentication as expressed in
Tinkerpop (https://tinkerpop.apache.org/docs/3.0.2-incubating/#_developing_a_driver)
documentation.

Creates a client that, on first request, it replies to the AUTHENTICATE
(407) response provided by the server, and keeps the socket for future
requests, not having to open one each time.